Add files via upload
Hey guys I uploaded all the code except for the dependencies in the Node Modules file. If you decide to download and view the code make sure you copy and paste node modules from an original react app. It should be exactly the same. Let me know if you do not understand anything.